First version of energy usage response from all capable devices known

+++ elis-energy-service/src/main/java/se/mah/elis/external/energy/beans/EnergyDataBean.java
@@ -10,6 +10,6 @@ public class EnergyDataBean {
 	public String timestamp;
 	
 	@XmlElement
-	public float kwh;
+	public double kwh;
 	
 }
